feat: implement Navbar component to enhance layout and navigation functionality
